Extracting dummy data to module
